+++ src/Basescape/TransferItemsState.h
@@ -67,7 +67,7 @@ private:
 	/// Gets type of selected item.
 	enum TransferType getType(size_t selected) const;
 	/// Gets item Index.
-	int getItemIndex(size_t selected) const;
+	size_t getItemIndex(size_t selected) const;
 public:
 	/// Creates the Transfer Items state.
 	TransferItemsState(Game *game, Base *baseFrom, Base *baseTo);
