+++ src/Engine/Game.h
@@ -71,7 +71,7 @@ public:
 	/// Sets the game's audio volume.
 	void setVolume(int sound, int music, int ui);
 	/// Adjusts a linear volume level to an exponential one.
-	static float volumeExponent(int volume);
+	static double volumeExponent(int volume);
 	/// Gets the game's display screen.
 	Screen *getScreen() const;
 	/// Gets the game's cursor.
