+++ src/Ruleset/SoldierNamePool.h
@@ -48,7 +48,7 @@ public:
 	/// Generates a new name from the pool.
 	std::wstring genName(SoldierGender *gender) const;
 	/// Generates an int representing the index of the soldier's look, when passed the maximum index value.
-	int genLook(size_t numLooks);
+	size_t genLook(size_t numLooks);
 };
 
 }
