+++ src/Savegame/AlienMission.h
@@ -47,10 +47,10 @@ class AlienMission
 private:
 	const RuleAlienMission &_rule;
 	std::string _region, _race;
-	unsigned _nextWave;
-	unsigned _nextUfoCounter;
-	unsigned _spawnCountdown;
-	unsigned _liveUfos;
+	size_t _nextWave;
+	size_t _nextUfoCounter;
+	size_t _spawnCountdown;
+	size_t _liveUfos;
 	int _uniqueID;
 	const AlienBase *_base;
 public:
@@ -75,9 +75,9 @@ public:
 	/// Sets the mission's race.
 	void setRace(const std::string &race) { _race = race; }
 	/// Gets the minutes until next wave spawns.
-	unsigned getWaveCountdown() const { return _spawnCountdown; }
+	size_t getWaveCountdown() const { return _spawnCountdown; }
 	/// Sets the minutes until next wave spawns.
-	void setWaveCountdown(unsigned minutes);
+	void setWaveCountdown(size_t minutes);
 	/// Sets the unique ID for this mission.
 	void setId(int id);
 	/// Gets the unique ID for this mission.
@@ -94,7 +94,7 @@ public:
 	/// Handle UFO spawning for the mission.
 	void think(Game &engine, const Globe &globe);
 	/// Initialize with values from rules.
-	void start(unsigned initialCount = 0);
+	void start(size_t initialCount = 0);
 	/// Increase number of live UFOs.
 	void increaseLiveUfos() { ++_liveUfos; }
 	/// Decrease number of live UFOs.
@@ -113,7 +113,7 @@ private:
 	/// Spawn an alien base
 	void spawnAlienBase(const Globe &globe, Game &engine);
 	/// Select a destination (lon/lat) based on the criteria of our trajectory and desired waypoint.
-	std::pair<double, double> getWaypoint(const UfoTrajectory &trajectory, const unsigned int nextWaypoint, const Globe &globe, const RuleRegion &region);
+	std::pair<double, double> getWaypoint(const UfoTrajectory &trajectory, const size_t nextWaypoint, const Globe &globe, const RuleRegion &region);
 };
 
 }
